Client: Updated objects
