--- savewise-client/src/common/objects/Transaction.ts
@@ -1,7 +1,0 @@
-export interface TransactionForm {
-    id?: number;
-    categoryId: number | undefined;
-    amount: number | undefined;
-    date: string;
-    description: string;
-}
