+++ savewise-client/src/common/objects/vault.ts
@@ -1,0 +1,6 @@
+export interface Vault {
+    id?: number;
+    name: string;
+    userId: number;
+    amount: number;
+}
