+++ savewise-client/src/components/Dashboard/Dashboard.tsx
@@ -10,8 +10,7 @@ import { CategoriesResponse, Category } from '../../common/objects/categories';
 import CircularProgress from '@material-ui/core/CircularProgress';
 import { GetCategoriesInput } from '../../services/category-service';
 import { CategoryTypesId } from '../../common/objects/CategoryTypesId';
-import { Transaction } from '../../common/objects/transactions';
-import { TransactionForm } from '../../common/objects/Transaction';
+import { Transaction, TransactionForm } from '../../common/objects/transactions';
 import Modal from "@material-ui/core/Modal";
 import MenuItem from "@material-ui/core/MenuItem";
 import TextField from "@material-ui/core/TextField";
